refactor(dm): split ConversationScreen into cohesive modules (#703)
ConversationScreen.tsx was 1455 lines (grandfathered over the 1000-line cap).
Rather than line-golf, separate it by responsibility so each piece does one
thing — and it drops to 796 lines, off the baseline entirely:

- ConversationScreen.styles.ts     — presentation (createStyles + style type)
- utils/conversationItems.ts       — pure DM/zap → display-item shaping
                                      (formatDayHeader, buildZapItems,
                                      buildConversationItems); now unit-tested
- components/ConversationMessageRow — the memoised FlatList row (day-header /
                                      zap card / MessageBubble), props reused
                                      from MessageBubble so they can't drift
- hooks/useConversationComposerActions — send/upload/share orchestration +
                                      in-flight flags, layered above the
                                      context's raw send methods

Behaviour is unchanged (verbatim moves, strongly-typed wiring). Verified on the
LP emulator: the 1:1 thread, day dividers, message bubbles, composer, and
attach panel all render and respond. Removes the ConversationScreen baseline
from scripts/check-file-size.sh, and adds a selector-based Maestro smoke flow.
